ahlanApi/requirements.txt
asgiref==3.8.1
attrs>=25.3.0
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
docx2pdf==0.1.8
drf-spectacular==0.28.0
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml>=5.3.1
packaging==24.2
Pillow>=11.1.0
PyJWT==2.9.0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
django-crontab==0.7.1
django-apscheduler==0.6.2
apscheduler==3.10.4
